Python/Microservicio_CRUD_USUARIOS/requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework_simplejwt
cryptography
django-cors-headers==3.5.0
django-extensions
gunicorn==20.0.4
Pillow==8.1.1
redis